% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/custom_ggplot2_themes.R
\name{theme_thesishPurple}
\alias{theme_thesishPurple}
\title{Thesish purple version theme for ggplot2 plots.}
\usage{
theme_thesishPurple(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Given reference size.}

\item{base_family}{Given font family.}
}
\value{
Prints or saves the given plot in a purple color scheme
}
\description{
\code{theme_thesishPurple} gives a ggplot2 plot a formal but elegant
color scheme to integrate it thesis env's.
}
\details{
This function is part of a family of theming functions for ggplot2 plots
which includes blue, green, red, grey and minimalist combinations..
}
\examples{
p <- ggplot(mtcars) + geom_point(aes(x = wt, y = mpg,
colour=factor(gear))) + facet_wrap(~am)

p
p + theme_thesishPurple()
}
\seealso{
\code{\link{theme_bluish}} for blue,
  \code{\link{theme_greenish}} for green, \code{\link{theme_redish}}
  for red, \code{\link{theme_greyish}} for grey, as well as \code{\link{theme_minimalish}}
  for a black&white minimalist plot.

Other custom ggplot2 theming functions: \code{\link{plot_variants}};
  \code{\link{theme_bluish}};
  \code{\link{theme_thesishBlue}};
  \code{\link{theme_thesishGreen}};
  \code{\link{theme_thesishRed}};
  \code{\link{theme_thesish}}
}

