% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_simulation_script.R
\name{three_way_power}
\alias{three_way_power}
\title{Power analysis for a three-way ANOVA.}
\usage{
three_way_power(N1, N2, N3, n_values, eff_sizes, alpha)
}
\arguments{
\item{N1,N2,N3}{Number of levels for each factor. Must be a numeric value.}

\item{n_values}{Number of replicates. Can be a numeric value or a numeric vector for
complex simulations.}

\item{eff_sizes}{Effect size. Can ve a numeric value or a numeric vector for
complex simulations.}

\item{alpha}{Confidence level.}
}
\value{
Prints in console the results indicating the power of analisis for
  the given n's and effect sizes, and returns a data.frame with all the
  information to make easy to produce a plot (optimized for ggplot2).
}
\description{
\code{three_way_power} calculates the power of a three-way ANOVA.
}
\details{
Preliminar (empirical) power analyisis. Non-centrality parameter (ncp) is
obtained from \url{http://stackoverflow.com/questions/2710609/power-analysis-in-r-for-two-way-anova}
(in a comment from Stéphane Laurent). Also, the power calculations are
obtained as explained in \url{https://stat.ethz.ch/pipermail/r-help/2011-April/273931.html}
The effect size values can be empirical or obtained as in
\url{http://www.uccs.edu/~lbecker/}.
}
\examples{
n_values <- c(3,4,5)
eff_sizes <- seq(.8,2.2,.2)
N1 <- 2
N2 <- 8
N3 <- 5
alpha <- .05

foo.power <- three_way_power(N1,N2,N3,n_values,eff_sizes,alpha)
head(foo.power)
}

